Avoid `getaddrinfo_a()` if resolver threads are not exhausted
The asynchronous `getaddrinfo_a()` function, while non-blocking,
comes with a hidden cost: it spawns a new thread for every
single query. This frequent creation and destruction of threads
is inefficient and can become a major bottleneck.

Check the number of active DNS queries against the pool of
available resolver threads. If there are idle threads ready to
work, we delegate new DNS lookups to them using the synchronous
`getaddrinfo()`. It reuses existing threads and completely
avoids the overhead of creating new ones.

The costly `getaddrinfo_a()` is now only used as a fallback when
the entire pool of resolver threads is exhausted. This leads to
a more robust, efficient, and scalable DNS resolution mechanism.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>