gwproxy: Introduce --dns-server and --raw-dns
This is the initial integration of the new DNS resolver feature that
does not rely on getaddrinfo(). Introduce two new options in order to
utilize it.

1) --dns-server to be filled with DNS servers, for now it only supports
   a single host. In the futures, it may accept multiple hosts
   separated by commas.

2) --raw-dns to be filled with either 0 or 1. 0 to disable, 1 to
   enable.

While in there, plug the new DNS resolver data structure into struct
gwp_wrk and initialize it.

Note that this feature is guarded with CONFIG_NEW_DNS_RESOLVER, which
is disabled by default as it's still experimental. A future patch will
add a configure option to enable it.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>